revert to standalone plain Rmd
* The bookdown report as standalone is buggy and struggles to knit with R3.5.1.
We have two options: either generate bookdown as a website (i.e. multiple html
files), or just revert to the old-style Rmd report with all results in a single
Rmd file (which is what we're doing here). We can come back to bookdown if we
decide to host reports on AWS (which, btw, is what happens on Trello anyhow).

* Also removed unused manta_vcf param, renamed af/vcf/sv params