Use MultiQC repack to maintain report consistency
The MultiQC code used in umccrise 1.2.1 is no longer available in the
original GitHub repo or anywhere else. The version of MultiQC used in
umccrise 1.2.1 contained changes specifically needed by umccrise. To
maintain consistency, the same version of MultiQC was needed and so the
MultiQC code from the umccrise 1.2.1 Docker image was repacked as a
Conda package. See github.com/umccr/MultiQC/tree/umccrise_bcbio_aws.