Use DRAGEN output prefix to find files
RGSM was previously used to find DRAGEN files however this only works by
virtue of the having the DRAGEN output file prefix set to the same
value. It is often, but not always, the case that the value of RGSM and
DRAGEN output prefix are identifical. Appropriate updates have been made
in umccr/NGS_Utils#3, aligning logic here as well.